% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{grepv}
\alias{grepv}
\title{A grep for variable names  Function}
\usage{
grepv(pattern = NULL, data = NULL)
}
\arguments{
\item{pattern}{Character string containing the search pattern.}

\item{data}{Data frame containing the variable names to search through.}
}
\description{
This function returns all the column names of a data frame containing the pattern provided.
}
\examples{
grepv("ing", attitude)
}
\keyword{grep}

